reducer を一部 skill reducer に移した
